code/content-gen/requirements.txt
google-genai>=0.2.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
